refactor: move url helpers to their own dir out of index
