+++ src/components/urls/stringsToGoogleSearchQ.js
@@ -1,0 +1,5 @@
+export default function stringsToGoogleSearchQ(strings = []) {
+  return strings
+    .map((str) => str.replace(/\s/g, '+'))
+    .join('+')
+}
